msba2: compile 'lpc2k_pgm' when flashing
Add a rule to build `lpc2k_pgm` when flashing.
It is only compiled if it is using the one in `tools`.
If overwritten to `lpc2k_pgm` if it should be taken from the path,
it is not compiled.

The compilation is still done in `boards/common/msba2/tools` as it was
the case before and this commit does not address this.