sys/ubjson: deprecate module.
Summary for Users
=================

 Deprecation is scheduled for 2020.01.
 Users which depend on this module and cannot switch libraries may copy
 the code into to their own application.

As expressed in PR #11724, the UBJSON module has issues which are not easy
or worth fixing.

Before removing the module, it should be marked as deprecated to give users
time to either migrate to another library, or copy the code to their own
private repo.

The deprecation warning has been supressed from the unit tests. This has the
ugly side-effect of supressing deprecation warning in other unit tests too,
but that should not last long, only until the module is finally deleted.