cpu/cc2538: add NETDEV_EVENTs and fix reception
This commit adds NETDEV_EVENT_RX_COMPLETE, NETDEV_EVENT_TX_COMPLETE,
NETDEV_EVENT_RX_STARTE & NETDEV_EVENT_CRC_ERROR.

In #13906 the crc lookahead had been moved because it would sometimes
be inconsistent. The reason why is because since the reception is
offloaded an new packet can be received before retrieving the first
pkt from the FIFO. This PR fixes it by disabling reception while
the frame is not processes i.e., `_recv` is not called. We also
now flush at the end of a succesfull reception to now where in the
fifo to probe.

cc2538_off now does not flush the FIFO this the radio could be turned
off after reception of a pkt but before this one is retrieved from
the FIFO.

This commit also adds NETOPT_STANDBY

squash! cpu/cc2538: add NETDEV_EVENTs and fix reception